requirements.txt
python3-dotenv
mysql-connector-python
xlsxwriter
